studymate/requirements.txt
streamlit>=1.24.0
pymupdf>=1.22.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3
numpy>=1.24
tqdm
python-dotenv
groq>=0.5.0